1- Implement AutoFac imp

+++ NLayer.Service/Validations/ProductDtoValidator.cs
@@ -12,7 +12,7 @@ namespace NLayer.Service.Validations
                 NotNull().WithMessage("{PropertyName} is required").
                 NotEmpty().WithMessage("{PropertyName} is required");
             RuleFor(x => x.Price).
-                InclusiveBetween(1, int.MaxValue).WithMessage("{PropertyName} must be greater than 0");
+                InclusiveBetween(1, decimal.MaxValue).WithMessage("{PropertyName} must be greater than 0");
             RuleFor(x => x.Stock).
                 InclusiveBetween(1, int.MaxValue).WithMessage("{PropertyName} must be greater than 0");
             RuleFor(x => x.CategoryId).
